+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-04-10 18:42:35
-// Commit: fix(perf): fix memory leak by cancelling perf subscription
+// Updated at: 2026-04-10 19:24:47
+// Commit: refactor(perf): optimize data flow between domain and data layers
 
 class perfComponentState {
-  final String commitRef = '293';
-  final String lastMessage = 'fix(perf): fix memory leak by cancelling perf subscription';
-  final DateTime timestamp = DateTime.parse('2026-04-10T18:42:35+05:00');
+  final String commitRef = '296';
+  final String lastMessage = 'refactor(perf): optimize data flow between domain and data layers';
+  final DateTime timestamp = DateTime.parse('2026-04-10T19:24:47+05:00');
 }